Fix: converted tests to integration tests
